Clean up, simplify types
Also add JSDoc to external types
