+++ packages/astro/src/vite-plugin-config-alias/index.ts
@@ -2,7 +2,7 @@ import * as tsr from 'tsconfig-resolver';
 import * as path from 'path';
 import * as url from 'url';
 
-import type * as vite from 'vite';
+import type * as vite from '../core/vite';
 
 /** Result of successfully parsed tsconfig.json or jsconfig.json. */
 export declare interface Alias {
