+++ packages/astro/src/vite-plugin-markdown/index.ts
@@ -1,5 +1,5 @@
 import type { Plugin } from '../core/vite';
-import type { AstroConfig } from '../@types/astro-core';
+import type { AstroConfig } from '../@types/astro';
 
 import esbuild from 'esbuild';
 import fs from 'fs';
